backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
fastapi==0.115.14
h11==0.16.0
h3==4.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
sniffio==1.3.1
starlette==0.46.2
timezonefinder==6.5.9
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
